requirements.txt
tensorflow	
numpy
open3d
sklearn
scipy
graph_nets
sonnet
python-igraph